Move `get_options()` to the `Settings` class
